moved details of candidate preview to their own file.
