+++ frontend/src/components/import/import.module.scss
@@ -64,7 +64,6 @@
     color: "inherit";
 
     &.inactive {
-        // color: #495057;
         color: #403b31;
     }
